Split function for wasm generation into separate files
