+++ src/token/list.rs
@@ -76,70 +76,3 @@ impl RecursiveAnalysisElements for ListBranch {
 }
 
 
-impl ListBranch{
-    /// indexの展開
-    pub fn generate_contents_wasm(&self) -> Result<String, GenerateError> {
-        let mut assembly_text = String::default();
-        match &self.contents[0] {
-            ExprElem::ItemElem(item_b) => {
-                assembly_text.push_str(&item_b.generate_wasm()?);
-            }
-            _=>{
-                return Err(GenerateError::Deverror);
-            }
-        }
-        Ok(assembly_text)
-    }
-
-    pub fn generate_name_wasm(&self) -> Result<String, GenerateError>
-    {
-        let mut assembly_text = String::default();
-            if let ExprElem::WordElem(word_b) = &*self.name {
-                if word_b.contents == MEMORY_SPACE_NAME {
-                    // 特別なケース、メモリに直接アクセスするための方法を提供する
-                    // ```
-                    // __mem[0] = 0;
-                    // ```
-                    assembly_text.push_str(&self.generate_contents_wasm()?);
-                } else {
-                    // 通常のケース
-                    // ```
-                    // a[0] = 0;
-                    // ```
-                    todo!()
-                }
-            } else {
-                // 呼び出しの対象が名前ではない場合
-                // ```
-                // lst[0][0]
-                // ^^^^^^
-                // ```
-                todo!()
-            }
-            Ok(assembly_text)
-    }
-}
-
-// Wasm_gen ---
-
-/// wasmで、メモリから値を取得する命令を記述する
-pub fn wasm_get_memory_value_gen(index:i32)-> Result<String, GenerateError>
-{
-    let mut assembly_text = String::default();
-
-    assembly_text.push_str(&format!("i32.const {}", index));
-    assembly_text.push_str("i32.load\n");
-
-    Ok(assembly_text)
-}
-
-///  wasmで、メモリに値を書く命令を記述する
-pub fn wasm_set_memory_value_gen(index:i32, r_assembly_text:&str)-> Result<String, GenerateError>
-{
-    let mut assembly_text = String::default();
-
-    assembly_text.push_str(&format!("i32.const {}", index));
-    assembly_text.push_str(r_assembly_text);
-    assembly_text.push_str("i32.store\n");
-    Ok(assembly_text)
-}
