+++ src/token/paren_block.rs
@@ -1,8 +1,6 @@
 use crate::abs::ast::*;
-use crate::abs::gen::Wasm_gen;
 use crate::errors::parser_errors::ParserError;
 use crate::parser::expr_parser::ExprParser;
-use crate::errors::generate_errors::GenerateError;
 
 /// #ParenBlockBranch
 /// `()`を使用したプログラムにおけるデータを格納するstruct
@@ -68,58 +66,4 @@ impl ASTAreaBranch<ExprElem> for ParenBlockBranch {
     }
 }
 
-
-impl Wasm_gen for ParenBlockBranch{
-    fn generate_wasm(&self) -> Result<String, crate::errors::generate_errors::GenerateError> {
-        let mut assembly_text = String::default();
-        match self.contents.len(){
-            0 => {
-                // pass
-                // カッコの中に何もない場合
-                // ()
-                // ただ、ここではは想定されていないエラーをキャッチする必要がある
-                // ```
-                // (()) / a
-                // ```
-            }
-
-            1 => {
-                match &self.contents[0]{
-                    // example `(1+a)`
-                    ExprElem::FuncElem(func_b) => {
-                        assembly_text.push_str(&func_b.generate_wasm()?);
-                    }
-
-                    // example `(a)`
-                    ExprElem::WordElem(word_b) => {
-                        if word_b.self_is_num()? {
-                            assembly_text.push_str(&format!("i32.const {}\n", word_b.contents));
-                        } else {
-                            // もし何らかの変数だった場合
-                            assembly_text.push_str(&format!("local.get ${}\n", word_b.contents));
-                        }
-                    }
-
-                    // `((a + 1))`
-                    ExprElem::ParenBlockElem(paren_b) => {
-                        assembly_text.push_str(&paren_b.generate_wasm()?);
-                    }
-
-                    _ => {
-                        // pass
-                        return Err(GenerateError::Deverror);
-                    }
-                }
-
-            }
-
-            _ => {
-                // 対応していない
-                // `(,,,)`的なシチュエーション
-                return Err(GenerateError::Deverror);
-            }
-        }
-        Ok(assembly_text)
-    }
-}
